test: 썸네일 받아오기 refactor: util directory -> src로 이동
