+++ src/public/util/alert.js
@@ -1,0 +1,15 @@
+function alertmove(path,msg){
+
+    let script = `<script type="text/javascript">
+
+    alert("${msg}")
+
+    location.href="${path}"
+
+    </script>`
+
+    return script
+
+}
+
+module.exports = {alertmove}
