+++ src/views/test/uploadTest.html
@@ -1,0 +1,14 @@
+<!DOCTYPE html>
+<html lang="en">
+<head>
+    <meta charset="UTF-8">
+    <title>Title</title>
+</head>
+<body>
+    <h1> img 썸네일 </h1>
+    {% for item in content %}
+    <img src = {}>
+    {% endfor %}
+</body>
+</html>
+<script type="text/javascript" src="/js/test/thumbnail.js"></script>
